+++ appPlugins/ToolPanelize.py
@@ -1369,6 +1369,8 @@ class PanelizeUI:
         self.separator_line.setFrameShadow(QtWidgets.QFrame.Shadow.Sunken)
         grid3.addWidget(self.separator_line, 10, 0, 1, 2)
 
+        # FCGridLayout.set_common_column_size([grid0, grid1, grid2, grid3], 0, FCLabel)
+
         # #############################################################################################################
         # Generate Panel Button
         # #############################################################################################################